requirements.txt
# Python dependencies for the framework
boto3>=1.26.0
pyyaml>=6.0
python-dateutil>=2.8.0
